Remove casts

+++ src/cinematic/CinematicTexture.cpp
@@ -307,9 +307,9 @@ void CinematicGrid::ReajustUV() {
 		
 		Vec2i size = tex->getStoredSize();
 		
-		if((size.x != (int)mat->bitmap.x) || (size.y != (int)mat->bitmap.y)) {
-			float dx = (.999999f * (float)mat->bitmap.x) / ((float)size.x);
-			float dy = (.999999f * (float)mat->bitmap.y) / ((float)size.y);
+		if(size.x != mat->bitmap.x || size.y != mat->bitmap.y) {
+			float dx = (.999999f * mat->bitmap.x) / size.x;
+			float dy = (.999999f * mat->bitmap.y) / size.y;
 
 			int nb2 = mat->nbvertexs;
